Getting help from T.O
i still have an issue with the idle because for some reason the legs and the body needs different animation clip which doesnt make sense at all because i need all the parts in one animation clip.
NOT FIXED THO waitin on T.O to help me, i sent him the sprite and scene so.